+++ src/sql/addresses.sql
@@ -1,10 +1,11 @@
 DROP TABLE IF EXISTS addresses;
 CREATE TABLE IF NOT EXISTS addresses (
-    action_index   BIGINT UNSIGNED NOT NULL, -- Unique action index
-    fee_preference BIGINT UNSIGNED,
-    require_memo   BIGINT UNSIGNED,
-    memo_id        BIGINT UNSIGNED,          -- id of record in index_memos table 
-    status_id      BIGINT UNSIGNED           -- id of record in index_statuses table
+    action_index         BIGINT UNSIGNED NOT NULL, -- Unique action index
+    fee_preference       BIGINT UNSIGNED,
+    require_memo         BIGINT UNSIGNED,
+    dispenser_preference BIGINT UNSIGNED,          -- 1=Only owner can open dispenser (default), 2=Anyone
+    memo_id              BIGINT UNSIGNED,          -- id of record in index_memos table
+    status_id            BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index   ON addresses (action_index);